♻️ refactor(todolist): rename dir
Change deps accordingly.
